prog

+++ paper/rust-prob/src/main.rs
@@ -1,6 +1,6 @@
 mod utils;
 mod probs;
-pub type U = usize;
+pub type U = u128;
 
 fn main() {
     println!("Hello, world!");
